refractor: merge animation widgets into single file
